Refactored job hierarchy with superclass BackgroundJob and implemented in-progress job records
Created AdminJob class for admin specific async tasks.
