+++ app/jobs/admin_job.rb
@@ -1,0 +1,2 @@
+class AdminJob < BackgroundJob
+end
