+++ app/models/completed_job.rb
@@ -1,2 +1,5 @@
+# CompletedJob is no longer an accurate name for this model as the table now stores in-progress jobs as well.
+# It has not been renamed yet to avoid associated headaches.
+
 class CompletedJob < ApplicationRecord
 end
